// projects/blog/src/app/types/post.type.ts
export type Post = {
  id: string;
  title: string;
  description: string;
  date: string;
  fileName: string;
  imageUrl: string;
};

export type PostSectionType = {
  tag: 'h1' | 'h2' | 'h3' | 'code' | 'li' | 'blockquote' | 'p' | 'a' | 'img';
  content: string | string[];
  meta?: {
    language?: string;
    alt?: string;
    href?: string;
    id?: string;
  };
};